requirements.txt
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.5.1
nltk==3.8.1
spacy==3.6.1
requests==2.31.0
beautifulsoup4==4.12.2
joblib==1.2.0
colorama==0.4.6
tqdm==4.65.0
newsapi-python==0.2.7
sentence-transformers>=2.2.2
newspaper3k==0.2.8
matplotlib==3.7.2
python-dotenv==1.0.0
